+++ src/gslc/semversion.fs
@@ -1,5 +1,5 @@
 ﻿module semversion
 
 let versionMajor = 0
-let versionMinor = 2
-let versionPatch = 2
+let versionMinor = 3
+let versionPatch = 0