+++ RELEASE_NOTES.md
@@ -1,2 +1,5 @@
+### 0.3.0
+  * FIX - uppercasing lowercase reference sequences on load
+  * extension of plugin mechanism to allow plugins to provide pragmas
 ### 0.2.2
   * initial public release of GSLC